fix gov balance maxxing at 2.1billion

+++ src-tauri/src/entities/business.rs
@@ -201,9 +201,9 @@ impl Business {
         owner.business_pay(self, owner_expected_income);
     }
 
-    pub fn pay_tax(&mut self, government_balance: &mut u128, amount: f32) {
+    pub fn pay_tax(&mut self, government_balance: &mut u64, amount: f32) {
         if amount < 0. { return }
         self.balance -= amount;
-        *government_balance += amount as u128;
+        *government_balance += amount as u64;
     }
 }
