+++ src-tauri/src/entities/person/person.rs
@@ -316,10 +316,10 @@ impl Person {
         payee.balance -= amount;
     }
     
-    pub fn pay_tax(&mut self, government_balance: &mut u128, amount: f32) {
+    pub fn pay_tax(&mut self, government_balance: &mut u64, amount: f32) {
         if amount < 0. { return }
         self.balance -= amount;
-        *government_balance += amount as u128;
+        *government_balance += amount as u64;
     }
 
     pub fn check_birthday(&mut self, date: &Date) {
